fix: update broken registry path references after folder restructure
Update imports, dynamic imports, tsconfig path alias, CI workflow
trigger, and AGENTS.md to point to docs/registry/react/ instead of
the old docs/registry/ flat structure.
